+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 
 - `:ws_url` option is now used without modifications WYSIWYG.
 - `Network.requestWillBeSent` callback didn't handle params in a type-safe way
+- `Page.frameStoppedLoading` callback shouldn't wait for document_node_id response
 
 ### Removed